pyproject.toml
[project]
name = "better-whoami"
version = "0.1.0-dev"
description = ""
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "litestar @ git+https://github.com/litestar-org/litestar.git@v3.0",
    "psutil (>=7.0.0,<8.0.0)",
    "dnspython (>=2.7.0,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "uvloop (>=0.21.0,<0.22.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.9"
pyright = "^1.1.400"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
  "C",
  "DTZ",
  "E",
  "F",
  "FLY",
  "G002",
  "G003",
  "G010",
  "G101",
  "I",
  "N",
  "PL",
  "Q",
  "RET",
  "RUF",
  "S",
  "T10",
  "T20",
  "UP",
  "W",
]
ignore = [
  "N805",
  "PLR0912",
  "PLR0913",
  "RUF100",
  "S101",
  "S104",
]
fixable = ["ALL"]

[tool.ruff.lint.pycodestyle]
max-line-length = 120
